refactor(domain): figuring out pub use to re export
