+++ service/domain/src/game/model/board.rs
@@ -1,0 +1,5 @@
+use crate::game::model::{Card, CardState};
+
+pub const BOARD_SIZE: usize = 25;
+pub type Board = [Card; BOARD_SIZE];
+pub type BoardState = [CardState; BOARD_SIZE];
