+++ service/domain/src/game/model/team.rs
@@ -1,0 +1,9 @@
+use std::fmt::Display;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Display, Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+pub enum Team {
+    Blue,
+    Red,
+}
